pyproject.toml
[project]
name = "grabette-gripper"
version = "0.1.0"
description = "gRPC motor+camera service for GRABETTE gripper on Pi Zero 2W"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "pillow>=12.1.1",
    "protobuf>=4.25.0",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
rpi = [
    "picamera2",
    "rustypot",
    "numpy<2",  # system simplejpeg is compiled against numpy 1.x
]
dev = [
    "grpcio-tools>=1.60.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
